+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Features
 * Make protocol messages pluggable (PYTHON-956)
 * Make protocol versions pluggable (PYTHON-954)
 * Make types pluggable (PYTHON-957)
+* Create a structure for protocol constants (PYTHON-960)
 * cqlengine: Remove default limit on QuerySets (PYTHON-517)
 * cqlengine: asynchronous execution support (PYTHON-605)
 * cqlengine: disallow Counter create, save operations (PYTHON-497)